Task1: add price increase scheduling (#4)
* first commit

* feature: реализован обычный шедулер
feature: добавлены условия инициализации для обычного шедулера

* feature: реализованы условия инициализации простого и оптимизированного шедулера
feature: части реализована аннотация трекающая время

* feature: реализована аннотация для трекинга времени

* feature: реализован оптимизированный шедулер без записи в файл

* feature: реализована запись результатов обновления в файл

* fix: исправлен тип данных price у сущности product

* feature: добавлена эксклюзивная блокировка для оптимизированного шедулера
fix: исправлено увеличение цены товара

* fix: добавлен перенос строки после имен полей

* feature: добавлено логирование в методы аспектов

* fix: исправлен подсчет страниц в обновляемом шедулере.
fix: разнесены скрипты инсертов таблиц

* fix: исправлен подсчет секунд в аспектах

* fix: исправлен подсчет секунд в аспектах

* fix: исправлены замечания

* fix: исправлены замечания

* feature: исправлены замечания

* fix: правки по ревью

* fix: правки по ревью(2)

+++ src/test/java/com/warehouse/myshop/product/ProductTestBase.java
@@ -9,6 +9,7 @@ import com.warehouse.myshop.product.mapper.ProductMapper;
 import com.warehouse.myshop.product.model.Product;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -59,7 +60,7 @@ public abstract class ProductTestBase {
                 .name("test")
                 .articleNumber("te-st-article")
                 .description("test description")
-                .price(50.5)
+                .price(BigDecimal.valueOf(50.5))
                 .quantity(5)
                 .build();
     }
@@ -70,7 +71,7 @@ public abstract class ProductTestBase {
                 .name("test")
                 .articleNumber("te-st-article")
                 .description("test description")
-                .price(50.5)
+                .price(BigDecimal.valueOf(50.5))
                 .quantity(10)
                 .build();
     }
